+++ src/main.rs
@@ -15,17 +15,9 @@ mod eval;
 mod object;
 mod read;
 mod repl;
+mod result;
 
 use repl::repl;
-
-use object::Object;
-use error::Error;
-
-#[doc(hidden)]
-pub type AthirResult = std::result::Result<Box<Object>, Error>;
-
-#[doc(hidden)]
-pub type AthirVecResult = std::result::Result<Vec<Box<Object>>, Error>;
 
 /// Main entry point
 ///
